Backend-da/requirements.txt

fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
clerk-backend-api
